move files to pages
