+++ src/renderer/src/pages/Create/index.tsx
@@ -6,12 +6,12 @@ function Create(): JSX.Element {
 
   return (
     <div>
-      <div className="text-blue-200">Other</div>
+      <div className="text-blue-200">Create page</div>
       <button
         className="bg-blue-600 hover:bg-blue-700 text-white text-sm px-4 py-2  border rounded-full"
         onClick={() => navigate('/')}
       >
-        Nav to home
+        Nav to Search
       </button>
     </div>
   )
